Added graphics

See merge request om.stu/omstu.oop/fit-191/do-iv!7

+++ include/components/positioncomponent.h
@@ -6,10 +6,10 @@
 class PositionComponent : public IComponent {
  private:
   friend class MovementSystem;
-  int x, y;
+  float x, y;
 
  public:
-  PositionComponent(const int _x, const int _y) : x(_x), y(_y) {}
-  int getX() const;
-  int getY() const;
+  PositionComponent(const float _x, const float _y) : x(_x), y(_y) {}
+  float getX() const;
+  float getY() const;
 };
